std::hash<short> -> std::hash<int64_t>

+++ test/tests/libcxx/unord.set/unord.set.cnstr/deduct.pass.cpp
@@ -78,37 +78,37 @@ void main()
 
     {
     const int arr[] = { 1, 2, 1, INT_MAX, 3 };
-    momo::stdish::unordered_set s(std::begin(arr), std::end(arr), 42, std::hash<short>());
+    momo::stdish::unordered_set s(std::begin(arr), std::end(arr), 42, std::hash<int64_t>());
 
-    ASSERT_SAME_TYPE(decltype(s), momo::stdish::unordered_set<int, std::hash<short>>);
+    ASSERT_SAME_TYPE(decltype(s), momo::stdish::unordered_set<int, std::hash<int64_t>>);
     assert(std::is_permutation(s.begin(), s.end(), std::begin(expected_s), std::end(expected_s)));
     }
 
     {
     const int arr[] = { 1, 2, 1, INT_MAX, 3 };
-    momo::stdish::unordered_set s(std::begin(arr), std::end(arr), 42, std::hash<short>(), test_allocator<int>(0, 40));
+    momo::stdish::unordered_set s(std::begin(arr), std::end(arr), 42, std::hash<int64_t>(), test_allocator<int>(0, 40));
 
-    ASSERT_SAME_TYPE(decltype(s), momo::stdish::unordered_set<int, std::hash<short>, std::equal_to<int>, test_allocator<int>>);
+    ASSERT_SAME_TYPE(decltype(s), momo::stdish::unordered_set<int, std::hash<int64_t>, std::equal_to<int>, test_allocator<int>>);
     assert(std::is_permutation(s.begin(), s.end(), std::begin(expected_s), std::end(expected_s)));
     assert(s.get_allocator().get_id() == 40);
     }
 
     {
-    momo::stdish::unordered_set<int, std::hash<short>, std::equal_to<>, test_allocator<int>> source;
+    momo::stdish::unordered_set<int, std::hash<int64_t>, std::equal_to<>, test_allocator<int>> source;
     momo::stdish::unordered_set s(source);
     ASSERT_SAME_TYPE(decltype(s), decltype(source));
     assert(s.size() == 0);
     }
 
     {
-    momo::stdish::unordered_set<int, std::hash<short>, std::equal_to<>, test_allocator<int>> source;
+    momo::stdish::unordered_set<int, std::hash<int64_t>, std::equal_to<>, test_allocator<int>> source;
     momo::stdish::unordered_set s{source};  // braces instead of parens
     ASSERT_SAME_TYPE(decltype(s), decltype(source));
     assert(s.size() == 0);
     }
 
     {
-    momo::stdish::unordered_set<int, std::hash<short>, std::equal_to<>, test_allocator<int>> source;
+    momo::stdish::unordered_set<int, std::hash<int64_t>, std::equal_to<>, test_allocator<int>> source;
     momo::stdish::unordered_set s(source, test_allocator<int>(0, 41));
     ASSERT_SAME_TYPE(decltype(s), decltype(source));
     assert(s.size() == 0);
@@ -116,7 +116,7 @@ void main()
     }
 
     {
-    momo::stdish::unordered_set<int, std::hash<short>, std::equal_to<>, test_allocator<int>> source;
+    momo::stdish::unordered_set<int, std::hash<int64_t>, std::equal_to<>, test_allocator<int>> source;
     momo::stdish::unordered_set s{source, test_allocator<int>(0, 42)};  // braces instead of parens
     ASSERT_SAME_TYPE(decltype(s), decltype(source));
     assert(s.size() == 0);
@@ -138,25 +138,25 @@ void main()
     }
 
     {
-    momo::stdish::unordered_set s({ 1, 2, 1, INT_MAX, 3 }, 42, std::hash<short>());
+    momo::stdish::unordered_set s({ 1, 2, 1, INT_MAX, 3 }, 42, std::hash<int64_t>());
 
-    ASSERT_SAME_TYPE(decltype(s), momo::stdish::unordered_set<int, std::hash<short>>);
+    ASSERT_SAME_TYPE(decltype(s), momo::stdish::unordered_set<int, std::hash<int64_t>>);
     assert(std::is_permutation(s.begin(), s.end(), std::begin(expected_s), std::end(expected_s)));
     }
 
 #ifdef LIBCPP_HAS_BAD_NEWS_FOR_MOMO
     {
-    momo::stdish::unordered_set s({ 1, 2, 1, INT_MAX, 3 }, 42, std::hash<short>(), std::equal_to<>());
+    momo::stdish::unordered_set s({ 1, 2, 1, INT_MAX, 3 }, 42, std::hash<int64_t>(), std::equal_to<>());
 
-    ASSERT_SAME_TYPE(decltype(s), momo::stdish::unordered_set<int, std::hash<short>, std::equal_to<>>);
+    ASSERT_SAME_TYPE(decltype(s), momo::stdish::unordered_set<int, std::hash<int64_t>, std::equal_to<>>);
     assert(std::is_permutation(s.begin(), s.end(), std::begin(expected_s), std::end(expected_s)));
     }
 #endif
 
     {
-    momo::stdish::unordered_set s({ 1, 2, 1, INT_MAX, 3 }, 42, std::hash<short>(), std::equal_to<>(), test_allocator<int>(0, 43));
+    momo::stdish::unordered_set s({ 1, 2, 1, INT_MAX, 3 }, 42, std::hash<int64_t>(), std::equal_to<>(), test_allocator<int>(0, 43));
 
-    ASSERT_SAME_TYPE(decltype(s), momo::stdish::unordered_set<int, std::hash<short>, std::equal_to<>, test_allocator<int>>);
+    ASSERT_SAME_TYPE(decltype(s), momo::stdish::unordered_set<int, std::hash<int64_t>, std::equal_to<>, test_allocator<int>>);
     assert(std::is_permutation(s.begin(), s.end(), std::begin(expected_s), std::end(expected_s)));
     assert(s.get_allocator().get_id() == 43);
     }
@@ -172,9 +172,9 @@ void main()
 
     {
     const int arr[] = { 1, 2, 1, INT_MAX, 3 };
-    momo::stdish::unordered_set s(std::begin(arr), std::end(arr), 42, std::hash<short>(), test_allocator<int>(0, 44));
+    momo::stdish::unordered_set s(std::begin(arr), std::end(arr), 42, std::hash<int64_t>(), test_allocator<int>(0, 44));
 
-    ASSERT_SAME_TYPE(decltype(s), momo::stdish::unordered_set<int, std::hash<short>, std::equal_to<int>, test_allocator<int>>);
+    ASSERT_SAME_TYPE(decltype(s), momo::stdish::unordered_set<int, std::hash<int64_t>, std::equal_to<int>, test_allocator<int>>);
     assert(std::is_permutation(s.begin(), s.end(), std::begin(expected_s), std::end(expected_s)));
     assert(s.get_allocator().get_id() == 44);
     }
@@ -188,9 +188,9 @@ void main()
     }
 
     {
-    momo::stdish::unordered_set s({ 1, 2, 1, INT_MAX, 3 }, 42, std::hash<short>(), test_allocator<int>(0, 42));
+    momo::stdish::unordered_set s({ 1, 2, 1, INT_MAX, 3 }, 42, std::hash<int64_t>(), test_allocator<int>(0, 42));
 
-    ASSERT_SAME_TYPE(decltype(s), momo::stdish::unordered_set<int, std::hash<short>, std::equal_to<int>, test_allocator<int>>);
+    ASSERT_SAME_TYPE(decltype(s), momo::stdish::unordered_set<int, std::hash<int64_t>, std::equal_to<int>, test_allocator<int>>);
     assert(std::is_permutation(s.begin(), s.end(), std::begin(expected_s), std::end(expected_s)));
     assert(s.get_allocator().get_id() == 42);
     }
